requirements.txt
# Expert Network Search Copilot — dependencies
# pip install -r requirements.txt

# ── LangGraph + LangChain (Text-to-SQL agent) ───────────────────────────────
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# ── FastAPI ──────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.8.0
pydantic-settings>=2.4.0

# ── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# ── Pinecone ─────────────────────────────────────────────────────────────────
pinecone>=5.0.0

# ── Embeddings ───────────────────────────────────────────────────────────────
openai>=1.40.0          # used for OpenAI-compatible embedding endpoint

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
